Core/GameEventMgr: Fixed compile warnings (uint16 to uint8 conversion)

+++ src/server/game/Events/GameEventMgr.cpp
@@ -186,11 +186,11 @@ void GameEventMgr::StopEvent(uint16 event_id, bool overwrite)
 
             SQLTransaction trans = CharacterDatabase.BeginTransaction();
             PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_ALL_GAME_EVENT_CONDITION_SAVE);
-            stmt->setUInt8(0, event_id);
+            stmt->setUInt8(0, uint8(event_id));
             trans->Append(stmt);
 
             stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_GAME_EVENT_SAVE);
-            stmt->setUInt8(0, event_id);
+            stmt->setUInt8(0, uint8(event_id));
             trans->Append(stmt);
 
             CharacterDatabase.CommitTransaction(trans);
@@ -1530,12 +1530,12 @@ void GameEventMgr::HandleQuestComplete(uint32 quest_id)
                 SQLTransaction trans = CharacterDatabase.BeginTransaction();
 
                 PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_GAME_EVENT_CONDITION_SAVE);
-                stmt->setUInt8(0, event_id);
+                stmt->setUInt8(0, uint8(event_id));
                 stmt->setUInt32(1, condition);
                 trans->Append(stmt);
 
                 stmt = CharacterDatabase.GetPreparedStatement(CHAR_ADD_GAME_EVENT_CONDITION_SAVE);
-                stmt->setUInt8(0, event_id);
+                stmt->setUInt8(0, uint8(event_id));
                 stmt->setUInt32(1, condition);
                 stmt->setFloat(2, citr->second.done);
                 trans->Append(stmt);
@@ -1575,11 +1575,11 @@ void GameEventMgr::SaveWorldEventStateToDB(uint16 event_id)
     SQLTransaction trans = CharacterDatabase.BeginTransaction();
 
     PreparedStatement* stmt = CharacterDatabase.GetPreparedStatement(CHAR_DEL_GAME_EVENT_SAVE);
-    stmt->setUInt8(0, event_id);
+    stmt->setUInt8(0, uint8(event_id));
     trans->Append(stmt);
 
     stmt = CharacterDatabase.GetPreparedStatement(CHAR_ADD_GAME_EVENT_SAVE);
-    stmt->setUInt8(0, event_id);
+    stmt->setUInt8(0, uint8(event_id));
     stmt->setUInt8(1, mGameEvent[event_id].state);
     stmt->setUInt32(2, mGameEvent[event_id].nextstart ? uint32(mGameEvent[event_id].nextstart) : 0);
     trans->Append(stmt);
